virtio-media: Phase 2 fix second gref overflow in MMAP path
The MMAP command path had the same u32 gref_start + gref_count overflow
as the IMPORT path (fixed in Phase 0). Compute the bounds check in
64-bit here too.

+++ hw/virtio/virtio-media.c
@@ -3574,9 +3574,9 @@ release_respond:
 
             gref_start = addr / VIRTIO_MEDIA_GREF_PAGE_SIZE;
             gref_count = DIV_ROUND_UP(len, VIRTIO_MEDIA_GREF_PAGE_SIZE);
-            needed = sizeof(resp) + gref_count * sizeof(uint32_t);
+            needed = sizeof(resp) + (size_t)gref_count * sizeof(uint32_t);
             if (in_len < needed ||
-                gref_start + gref_count > s->gref_count) {
+                (uint64_t)gref_start + gref_count > s->gref_count) {
                 vmedia_write_resp_header(&resp.hdr, ENOSPC);
                 vmedia_iov_write(elem->in_sg, elem->in_num, 0,
                                  &resp, sizeof(resp));
